Fixes some bags for Teacher role

+++ src/main/java/com/schedule/proj/repository/TeacherRepository.java
@@ -1,6 +1,7 @@
 package com.schedule.proj.repository;
 
 
+import com.schedule.proj.model.Student;
 import com.schedule.proj.model.Subject;
 import com.schedule.proj.model.Teacher;
 import org.apache.tomcat.jni.User;
@@ -14,5 +15,5 @@ import java.util.Set;
 
 @Repository
 public interface TeacherRepository extends JpaRepository<Teacher, Integer> {
-    Teacher getByUserId(Long id);
+    Teacher getByUserId(int id);
 }
